+++ packages/scrabble-solver/src/components/NavButtons/index.ts
@@ -1,0 +1,1 @@
+export { default } from './NavButtons';
