+++ packages/scrabble-solver/src/components/index.ts
@@ -7,6 +7,7 @@ export { default as Key } from './Key';
 export { default as KeyMap } from './KeyMap';
 export { default as Loading } from './Loading';
 export { default as Logo } from './Logo';
+export { default as NavButtons } from './NavButtons';
 export { default as PlainTiles } from './PlainTiles';
 export { default as Radio } from './Radio';
 export { default as Results } from './Results';
