+++ packages/scrabble-solver/src/pages/index.module.scss
@@ -31,14 +31,6 @@
   user-select: none;
 }
 
-.iconButton {
-  margin-right: var(--spacing--l);
-
-  &:last-child {
-    margin-right: 0;
-  }
-}
-
 .contentWrapper {
   $board-size: 15;
   $tile-size-max: 60px;
